+++ CHANGES.rst
@@ -7,8 +7,30 @@ Changelog
 
 - Fix Plone6 compatibility.
   [cekk]
+
 - Removed unused type PrenotazioniFolderContainer.
   [cekk]
+
+- Added endpoint to get booking schema.
+  [daniele]
+
+- Avoid change gate, booking date, booking end from /edit;
+  this would allow you to skip the checks;
+  Fix profile registration name;
+  [lucabel]
+
+- Add @bookings endpoint to get booking items for a user
+  [foxtrot-dfm1]
+
+- Add a new endpoint to get booking details. (#40442).
+  [daniele]
+  
+- Add autoconfirm content rule to profile.
+  [foxtrot-dfm1]
+
+- Added field "cosa_serve" (#40445).
+  [daniele]
+
 
 1.7.1 (2023-03-28)
 ------------------